refactor: Introduce Physical Profile onboarding step
This commit refactors the onboarding flow by introducing a new step for "Physical Profile". This step will be used to collect the user's year of birth, weight, height, and preferred measurement system.

The "About Yourself" component has been renamed to "TrainingLevel" and is now responsible solely for selecting the user's training level. The layout of the training level selection buttons has also been updated.

New data classes and sealed interfaces have been added to represent length and weight with different units, as well as a `MeasurementSystem` enum and corresponding update events for the physical profile.

A utility function `toStringWithPrecision` has been added to format numbers to a specific decimal precision, which will be useful for displaying weight and height values.